Revert "Return the whole map bounds instead of our size for _getExpandedVisibleBounds so they are always valid"
This reverts commit ba0c2146cac507a7592a7ddc827e18b662514bda.

When using removeOutsideVisibleBounds the code had been returning
the current map view bounds from _map.getBounds(). This is a
smaller view than the _getExpandedVisibleBounds() code would
normally return, and cuts off markers outside the current map
view.

This is normally desired, but in an alternate projection or with
markers just on the periphery they are cut from the map's current
view bounding box.

Instead when disabling visible bounds removal, the bounds used to
cut markers and clusters should either be the whole world or the
bounds of all markers in the cluster. That is why I am reverting
the commit.